Move bun-orm, move models add DEBUG env
Move from sqlx to bun to make handling the database simpler.

During this process, the models have to be moved into a common pkg
to prevent import cycles.

They are re-exported in their own directories to make them more
discoverable.

The environment variable "DEBUG" was added to enable verbose logging.